requirements.txt
apache-airflow>=2.7.1
apache-airflow-providers-postgres>=5.7.1
apache-airflow-providers-common-sql>=1.9.0
cryptography>=42.0.2
certifi>=2024.2.2
urllib3>=2.2.0
python-dotenv>=1.0.0

# Test dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-env>=1.1.1
pytest-timeout>=2.2.0
sqlparse>=0.4.4
